fix: api refactoring
